requirements.txt
fastapi==0.112.2
tortoise-orm==0.21.6
aiogram==3.12.0
uvicorn==0.30.6
python-dotenv==1.0.1
asyncpg==0.29.0
